\name{evaluate-methods}
\docType{methods}
\alias{evaluate-methods}
\alias{evaluate}
\title{Generic evaluate method}
\description{
 Methods for function \code{evaluate}.
}
\section{Methods}{
\describe{
\item{object = "SplineBasis", x = "numeric"}{ Evaluates a \code{SplineBasis} object for the spline basis curves at points \code{x}. See \linkS4class{SplineBasis}}
}}
\keyword{methods}
