\name{integrate-methods}
\docType{methods}
\alias{integrate-methods}
\alias{integrate}
\title{ Methods for Function integrate}
\description{
Methods for function \code{integrate}.  \code{integrate} integrates generic objects for which an integral is defined.
}
\section{Methods}{
\describe{
	\item{object = "SplineBasis"}{ Returns a new \code{SplineBasis} object for the integral of the basis functions.  See \code{\linkS4class{SplineBasis}}}
}}
\keyword{methods}
\keyword{math}
