\name{orthogonalize-methods}
\docType{methods}
\alias{orthogonalize-methods}
\alias{orthogonalize}
\title{ Methods for Function orthogonalize}
\description{
 A generic function for orthogonalizing an object and returning the orthogonal object
}
\section{Methods}{
\describe{

\item{object = "SplineBasis"}{ Orthogonalize the spline basis functions. See \code{\link{SplineBasis}}}
}}
\keyword{methods}
