\name{orthogonalsplinebasis-package}
\alias{orthogonalsplinebasis-package}
\alias{orthogonalsplinebasis}
\docType{package}
\title{A Matrix Representation for Spline Basis Functions}
\description{This package provides functions for manipulation of spline basis functions.
A matrix representation for the basis functions is at the center of the functions.
The matrix representation simplifies the process of orthogonalization as well as differentiation and integration.
}
\author{
	Andrew Redd
	Maintainer: Andrew Redd \email{andrew.redd@hsc.utah.edu}
}
\keyword{package}
