\name{OrthogonalizeBasis}
\alias{OrthogonalizeBasis}
\title{ Orthogonalize a Spline Basis }
\description{
  Specific function for orthogonalizing the functions in a  \code{SplineBasis} object.
}
\usage{
OrthogonalizeBasis(object, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ A \code{SplineBasis} object}
  \item{\dots}{ ignored }
}
\value{
  An \code{\linkS4class{OrthogonalSplineBasis}} object.
}
\seealso{ \code{\linkS4class{OrthogonalSplineBasis}}, \code{\linkS4class{SplineBasis}},\code{\link{orthogonalize}} }
\keyword{ classes }
