\name{orthogonalsplinebasis-package}
\alias{orthogonalsplinebasis-package}
\alias{orthogonalsplinebasis}
\docType{package}
\title{A Matrix Representation for Spline Basis Functions}
\description{This package provides functions for manipluation of spline basis functions.  
A matrix representation for the basis fucntions is at the center of the functions. 
The matrix representation simplifies the process of orthogonalization as well as differention and integration.
}
\author{
	Andrew Redd
	Maintainer: Andrew Redd <aredd at stat.tamu.edu>
}
\keyword{package}